src/requirements.txt
asgiref==3.8.1
Django==4.2
sqlparse==0.5.2
djangorestframework==3.15.2
djoser==2.3.1
psycopg2-binary==2.9.10
gunicorn==23.0.0